backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
PyMuPDF
beautifulsoup4
requests
groq
python-dotenv
httpx
aiofiles
